+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased changes
 * ``kp.tools.AnyBar`` added to control the AnyBar macOS app, including a
   pipeline integration: ``kp.Pipeline(anybar=True)``
 * ``km3pipe runtable`` is now a standalone command line tool: ``runtable``
+* ``km3pipe runinfo`` is now a standalone command line tool: ``runinfo``
 
 
 Version 7